Change double to float

+++ Infrastructure/AI/ForecastOutput.cs
@@ -4,8 +4,8 @@
     {
         public double MAE { get; set; }
         public double RMSE { get; set; }
-        public double[] ForecastedValues { get; set; }
-        public double[] LowerBounds { get; set; }
-        public double[] UpperBounds { get; set; }
+        public float[] ForecastedValues { get; set; }
+        public float[] LowerBounds { get; set; }
+        public float[] UpperBounds { get; set; }
     }
 }
